vk_functions: Prefix functions with vkfw_
Previously the shared object build was broken, as when dlsym was
executed, it returned the symbols in the shared object. This did not
affect the static build, because symbols in the static build are not
exported.

Rename all Vulkan functions to vkfw_vkSomething instead of vkSomething
and use defines to avoid the broken behavior.

This is a breaking change, hence major version bump.