wayland: Initial implementation.
The Wayland backend is currently in a very incomplete state. It can
successfully render "Hello, Triangle!", but it cannot handle window
events yet. Also, it probably only works on my machine.

The Wayland event dispatch model will require us to break applications
by replacing the "one event at a time"-style event handling with a more
traditional model of "call this function once, and everything is
polled". XCB and Win32 backends will be undisturbed by this, as it can
be implemented by adding a new PFN to VKFWwindowapi_T and implementing
the new API on top of the old one if it is nullptr.